Conflicts:
	scilab/Scilab.sln
	scilab/Scilab_f2c.sln

+++ scilab/modules/string/sci_gateway/c/sci_length.c
@@ -28,7 +28,6 @@ length('123')  is  3 .  length([1,2;3,4])  is  4 .                     */
 #include <string.h>
 #include <stdio.h>
 #include "gw_string.h"
-#include "stack-c.h"
 #include "api_scilab.h"
 #include "core_math.h"
 #include "MALLOC.h"
@@ -366,7 +365,7 @@ static int lengthMList(const char *fname, int *piAddressVar)
                 if (isScilabFunction(overloadFunctionName))
                 {
                     int lw = 1 + Top - Rhs;
-                    C2F(overload)(&lw, (char*)fname, strlen(fname));
+                    C2F(overload)(&lw, (char*)fname, (unsigned long)strlen(fname));
                     return 0;
                 }
                 else
